+++ keystorage/src/main/java/tech/pegasys/web3signer/keystorage/azure/AzureKeyVault.java
@@ -15,6 +15,9 @@ package tech.pegasys.web3signer.keystorage.azure;
 import tech.pegasys.web3signer.keystorage.common.MappedResults;
 import tech.pegasys.web3signer.keystorage.common.SecretValueMapperUtil;
 
+import java.net.URI;
+import java.net.http.HttpRequest;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
@@ -24,31 +27,41 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 
+import com.azure.core.credential.AccessToken;
 import com.azure.core.credential.TokenCredential;
+import com.azure.core.credential.TokenRequestContext;
 import com.azure.core.exception.ResourceNotFoundException;
 import com.azure.core.http.rest.PagedIterable;
 import com.azure.identity.ClientSecretCredentialBuilder;
 import com.azure.identity.ManagedIdentityCredentialBuilder;
 import com.azure.security.keyvault.keys.KeyClient;
 import com.azure.security.keyvault.keys.KeyClientBuilder;
+import com.azure.security.keyvault.keys.KeyServiceVersion;
 import com.azure.security.keyvault.keys.cryptography.CryptographyClient;
 import com.azure.security.keyvault.keys.cryptography.CryptographyClientBuilder;
+import com.azure.security.keyvault.keys.cryptography.models.SignatureAlgorithm;
 import com.azure.security.keyvault.keys.models.KeyProperties;
 import com.azure.security.keyvault.keys.models.KeyVaultKey;
 import com.azure.security.keyvault.secrets.SecretClient;
 import com.azure.security.keyvault.secrets.SecretClientBuilder;
 import com.azure.security.keyvault.secrets.models.KeyVaultSecret;
 import com.azure.security.keyvault.secrets.models.SecretProperties;
+import io.vertx.core.json.JsonObject;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.apache.tuweni.bytes.Bytes;
 
 public class AzureKeyVault {
 
   private static final Logger LOG = LogManager.getLogger();
-
   private final TokenCredential tokenCredential;
   private final SecretClient secretClient;
   private final KeyClient keyClient;
+  private static final List<String> SCOPE = List.of("https://vault.azure.net/.default");
+  private final TokenRequestContext tokenRequestContext =
+      new TokenRequestContext().setScopes(SCOPE);
+
+  private Optional<AccessToken> maybeToken = Optional.empty();
 
   public static AzureKeyVault createUsingClientSecretCredentials(
       final String clientId,
@@ -63,7 +76,6 @@ public class AzureKeyVault {
             .tenantId(tenantId)
             .executorService(executorService)
             .build();
-
     return new AzureKeyVault(tokenCredential, vaultName);
   }
 
@@ -101,6 +113,42 @@ public class AzureKeyVault {
         .credential(tokenCredential)
         .keyIdentifier(keyId)
         .buildClient();
+  }
+
+  public HttpRequest getRemoteSigningHttpRequest(
+      final byte[] data,
+      final SignatureAlgorithm signingAlgo,
+      final String vaultName,
+      final String azureKeyName,
+      final String azureKeyVersion) {
+
+    final String apiVersion = KeyServiceVersion.getLatest().getVersion();
+
+    final JsonObject jsonBody = new JsonObject();
+    jsonBody.put("alg", signingAlgo);
+    jsonBody.put("value", Bytes.of(data).toBase64String());
+
+    final String uriString =
+        constructAzureSignApiUri(vaultName, azureKeyName, azureKeyVersion, apiVersion);
+
+    final HttpRequest httpRequest =
+        HttpRequest.newBuilder(URI.create(uriString))
+            .header("Content-Type", "application/json")
+            .header("Authorization", "Bearer " + getOrRequestNewToken())
+            .POST(HttpRequest.BodyPublishers.ofString(jsonBody.toString()))
+            .build();
+
+    return httpRequest;
+  }
+
+  private String constructAzureSignApiUri(
+      final String keyVaultName,
+      final String keyName,
+      final String keyVersion,
+      final String apiVersion) {
+    return String.format(
+        "https://%s.vault.azure.net/keys/%s/%s/sign?api-version=%s",
+        keyVaultName, keyName, keyVersion, apiVersion);
   }
 
   public static String constructAzureKeyVaultUrl(final String keyVaultName) {
@@ -205,5 +253,13 @@ public class AzureKeyVault {
 
     return keyProperties.getTags() != null // return false if remote secret doesn't have any tags
         && keyProperties.getTags().entrySet().containsAll(tags.entrySet());
+  }
+
+  private String getOrRequestNewToken() {
+    if (maybeToken.isEmpty() || maybeToken.get().isExpired()) {
+      maybeToken = Optional.of(tokenCredential.getTokenSync(tokenRequestContext));
+    }
+
+    return maybeToken.get().getToken();
   }
 }